// src/app/shared/models/shared/tabla-maestra-relacion.ts
export class TablaMaestraRelacion {
  idTablaMaestraRelacion: number;
  idTabla_1: number;
  idColumna_1: number;
  valor_1: string;
  descripcion_1: string;
  idTabla_2: number;
  idColumna_2: number;
  valor_2: string;
  descripcion_2: string;
  idTipoRelacion_1: number;
  idTipoRelacion_2: number;

  editado: boolean;
}